+++ plutarch-ledger-api/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## 3.4.0 -- 29-04-2025
+
+### Changed
+
+* Bump Plutus dependencies
+
 ## 3.3.0 -- 01-29-2025
 
 ### Added