+++ plutarch-orphanage/CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## 1.2.0
+
+### Added
+
+* `Arbitrary`, `CoArbitrary` and `Function` instances for `MintValue` from
+  `plutus-ledger-api`'s `V3` modules. This mirrors the old behaviour of our
+  `MintValue` modifier.
+
+### Changed
+
+* Bump Plutus dependencies
+
 ## 1.1.0
 
 ### Added